+++ CHANGELOG.md
@@ -10,6 +10,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 To see unreleased changes, please see the [CHANGELOG on the master branch](https://github.com/gufolabs/gufo_acme/blob/master/CHANGELOG.md) guide.
 
+## [Unreleased]
+
+## Fixed
+
+* Fixed typo in exception class name.
+
 ## 0.1.1 - 2023-11-16
 
 ### Fixed